Move helpers to separate directory 'api_cmds'
